Fixed use of schema for promo validation/calc.
Apparently I'm insane, because I *swear* for *hours* this was the only
way I could get the plugin and API to talk to one another successfully,
but then smnirven got peeved and I took one more look at it, and I'll be
goddamned if it doesn't work just fine - just like he said.

I'm confused.  But that's nothing new.